Revert "refactor(command-manager-lib): change command-manager root dir to remote-command-manager"
This reverts commit 1123d87347c4b8287741ec109192523935bb85f4.
