+++ src/CMakeLists.txt
@@ -1,7 +1,7 @@
 project(message-manager CXX)
 
 add_subdirectory("${PROJECT_SOURCE_DIR}/communications-layer")
-add_subdirectory("${PROJECT_SOURCE_DIR}/remote-command-manager")
+add_subdirectory("${PROJECT_SOURCE_DIR}/command-manager")
 add_subdirectory("${PROJECT_SOURCE_DIR}/interactive-console")
 add_subdirectory("${PROJECT_SOURCE_DIR}/interactive-commands")
 add_subdirectory("${PROJECT_SOURCE_DIR}/arguments-parser")
