+++ src/command-manager/src/command-manager.cpp
@@ -1,4 +1,4 @@
-#include <remote-command-manager/command-manager.hpp>
+#include <command-manager/command-manager.hpp>
 #include <cstring>
 
 command_manager::command_manager(communications_layer_interface *communications_layer)
