+++ src/command-manager/test/CMakeLists.txt
